+++ backend/controllers/authController.js
@@ -1,0 +1,162 @@
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+const User = require("../models/User");
+
+async function signup(req, res) {
+  try {
+    const { firstName, lastName, email, password, confirmPassword } = req.body;
+
+    if (!firstName || !lastName || !email || !password || !confirmPassword) {
+      return res.status(403).send({
+        success: false,
+        message: "All Fields are required",
+      });
+    }
+
+    if (!email.includes("@gmail.com")) {
+      return res.status(400).send({
+        success: false,
+        message: "Invalid Email!",
+      });
+    }
+
+    if (password !== confirmPassword) {
+      return res.status(400).json({
+        success: false,
+        message:
+          "Password and Confirm Password do not match. Please try again.",
+      });
+    }
+
+    const existingUser = await User.findOne({ email });
+    if (existingUser) {
+      return res.status(400).json({
+        success: false,
+        message: "User already exists. Please login to continue.",
+      });
+    }
+
+    const hashedPassword = await bcrypt.hash(password, 10);
+
+    const user = await User.create({
+      firstName,
+      lastName,
+      email,
+      password: hashedPassword,
+      image: `https://api.dicebear.com/5.x/initials/svg?seed=${firstName} ${lastName}`,
+    });
+
+    return res.status(200).json({
+      success: true,
+      user,
+      message: "User registered successfully!",
+    });
+  } catch (error) {
+    console.log(error.message);
+    return res.status(500).json({
+      success: false,
+      message: "User cannot be registered, Please try again!",
+    });
+  }
+}
+
+async function login(req, res) {
+  try {
+    const { email, password } = req.body;
+
+    if (!email || !password) {
+      return res.status(400).json({
+        success: false,
+        message: `All fields are required!`,
+      });
+    }
+
+    const user = await User.findOne({ email });
+
+    if (!user) {
+      return res.status(401).json({
+        success: false,
+        message: `User is not Registered with Us. Please SignUp to Continue.`,
+      });
+    }
+
+    if (await bcrypt.compare(password, user.password)) {
+      const token = jwt.sign(
+        {
+          email: user.email,
+          id: user._id,
+        },
+        process.env.JWT_SECRET,
+        {
+          expiresIn: "24h",
+        }
+      );
+
+      user._doc.token = token;
+      user._doc.password = undefined;
+
+      const options = {
+        expires: new Date(Date.now() + 3 * 24 * 60 * 60 * 1000),
+        httpOnly: true,
+      };
+
+      res.cookie("token", token, options).status(200).json({
+        success: true,
+        token,
+        user,
+        message: `User Login Successfully!`,
+      });
+    } else {
+      return res.status(401).json({
+        success: false,
+        message: `Password is incorrect`,
+      });
+    }
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({
+      success: false,
+      message: `Login Failure, Please Try Again!`,
+    });
+  }
+}
+
+async function logout(req, res) {
+    try {
+        console.log("req.user:", req.user);
+        await User.findByIdAndUpdate(
+          req.user._id,
+          {
+            $unset: {
+              token: 1,
+            },
+          },
+          {
+            new: true,
+          }
+        );
+    
+        const options = {
+          httpOnly: true,
+          secure: true,
+        };
+    
+        res
+          .status(200)
+          .clearCookie("token", options)
+          .json({ message: "User logged out successfully" });
+      } catch (error) {
+        res.status(500).json({ error: "An error occurred during logout" });
+      }
+}
+
+function googleCallback(req, res) {
+  res.redirect("/profile");
+}
+
+module.exports = {
+  signup,
+  login,
+  logout,
+  googleCallback,
+};
